Break up long tests
